script.ts testado todas as funções rodaram, testei o APAGAR somente no User, funcionou 100% nas demais não fiz, mas é igual... testei o Criar, Editar e Ler de todos os models/type. Falta fazer agora a parte do servicos e rota em separado...

+++ API/prisma/migrations/20240203030404_init/migration.sql
@@ -3,7 +3,7 @@ CREATE TABLE "User" (
     "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
     "name" TEXT NOT NULL,
     "email" TEXT NOT NULL,
-    "telephone" INTEGER NOT NULL,
+    "telephone" BIGINT NOT NULL,
     "password" TEXT NOT NULL,
     "address" TEXT NOT NULL,
     "birthday" DATETIME NOT NULL,
